@@ -11,9 +11,9 @@ repositories {
 
 dependencies {
     // netty-bom
-    api(platform("io.netty:netty-bom:4.1.85.Final"))
+    api(platform("io.netty:netty-bom:4.1.87.Final"))
     // junit-bom
-    testImplementation(platform("org.junit:junit-bom:5.9.1"))
+    testImplementation(platform("org.junit:junit-bom:5.9.2"))
     // jackson2-bom
     api(platform("com.fasterxml.jackson:jackson-bom:2.14.1"))
     // rocketmq
@@ -24,13 +24,13 @@ dependencies {
     api(platform("io.prometheus:simpleclient_bom:0.16.0"))
 
     constraints {
-        implementation("org.slf4j:slf4j-api:1.7.36")
-        implementation("ch.qos.logback:logback-classic:1.2.11")
+        implementation("org.slf4j:slf4j-api:2.0.6")
+        implementation("ch.qos.logback:logback-classic:1.4.5")
         api("io.lettuce:lettuce-core:6.2.2.RELEASE")
         api("com.dslplatform:dsl-json-java8:1.9.9")
         api("com.jsoniter:jsoniter:0.9.23")
         api("com.aliyun.openservices:ons-client:2.0.4.Final")
-        val mongodbVersion = "4.8.1"
+        val mongodbVersion = "4.8.2"
         api("org.mongodb:bson:$mongodbVersion")
         api("org.mongodb:mongodb-driver-core:$mongodbVersion")
         api("org.mongodb:mongodb-driver-sync:$mongodbVersion")
@@ -55,13 +55,13 @@ java {
     withSourcesJar()
     withJavadocJar()
     toolchain {
-        languageVersion.set(JavaLanguageVersion.of(11))
+        languageVersion.set(JavaLanguageVersion.of(17))
     }
 }
 
 tasks.compileJava {
     options.encoding = "UTF-8"
-    options.release.set(11)
+    options.release.set(17)
 }
 
 tasks.javadoc {
